requirements.txt
# Core dependencies
python-telegram-bot[job-queue]>=20.0
pydantic>=2.0
cryptography>=41.0.0
langchain>=0.1.0
langchain-openai>=0.0.3
openai>=1.6.1
python-dotenv>=1.0.0
urllib3<2.0.0

# Database and storage
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
supabase>=2.0.0
psycopg2-binary>=2.9.9

# Async utilities
aiohttp>=3.9.0
asyncio>=3.4.3
web3>=6.11.3
eth-account>=0.9.0
eth-utils>=2.3.1
eth-typing>=3.5.1
eth-hash>=0.5.2
eth-abi>=4.2.1

# Monitoring and logging
psutil>=5.9.6
prometheus-client>=0.19.0
structlog>=23.2.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.2
pytest-cov>=4.1.0
pytest-mock>=3.10.0

# Development
black>=23.11.0
isort>=5.12.0
mypy>=1.7.1
ruff>=0.1.6
flake8>=6.0.0

# Type checking
types-python-dateutil>=2.8.2
types-requests>=2.31.0

# Documentation
mkdocs>=1.4.0
mkdocs-material>=9.1.0
mkdocstrings>=0.22.0

# Added dependencies
requests>=2.31.0
aiofiles>=23.2.1
aiodns>=3.1.1
cchardet>=2.1.7
ujson>=5.9.0 